# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube Viral Intelligence Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /titles/analyze:
    post:
      operationId: analyzeTitles
      tags: [titles]
      summary: Analyze competitor titles with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeTitlesBody"
      responses:
        "200":
          description: Title analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleAnalysisResponse"
...
  /scripts/analyze:
    post:
      operationId: analyzeScript
      tags: [scripts]
      summary: Analyze a competitor script with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeScriptBody"
      responses:
        "200":
          description: Script analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptAnalysisResponse"
...
components:
  schemas:
...
    AnalyzeTitlesBody:
      type: object
      properties:
        titles:
          type: array
          items:
            type: string
        outputMode:
          type: string
          enum: [docs, text]
      required: [titles, outputMode]
    TitleAnalysisResponse:
      type: object
      properties:
        id:
          type: integer
        analysis:
          type: string
        outputMode:
          type: string
        createdAt:
          type: string
      required: [id, analysis, outputMode, createdAt]
...
    AnalyzeScriptBody:
      type: object
      properties:
        script:
          type: string
        outputMode:
          type: string
          enum: [docs, text]
      required: [script, outputMode]
    ScriptAnalysisResponse:
      type: object
      properties:
        id:
          type: integer
        analysis:
          type: string
        outputMode:
          type: string
        createdAt:
          type: string
      required: [id, analysis, outputMode, createdAt]
...