# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube Viral Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Fetch latest videos from competitor channels
 */
export const fetchCompetitorVideosBodyCompetitorsMax = 3;

export const FetchCompetitorVideosBody = zod.object({
  competitors: zod
    .array(zod.string())
    .min(1)
    .max(fetchCompetitorVideosBodyCompetitorsMax),
  youtubeApiKey: zod.string().nullish(),
});

export const FetchCompetitorVideosResponse = zod.object({
  videos: zod.array(
    zod.object({
      id: zod.number(),
      competitor: zod.string(),
      title: zod.string(),
      views: zod.string(),
      url: zod.string(),
      publishedAt: zod.string().nullish(),
      thumbnailUrl: zod.string().nullish(),
    }),
  ),
  usedMockData: zod.boolean(),
});

/**
 * @summary Get all saved competitor videos
 */
export const GetVideosQueryParams = zod.object({
  competitor: zod.coerce.string().optional(),
});

export const GetVideosResponse = zod.object({
  videos: zod.array(
    zod.object({
      id: zod.number(),
      competitor: zod.string(),
      title: zod.string(),
      views: zod.string(),
      url: zod.string(),
      publishedAt: zod.string().nullish(),
      thumbnailUrl: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Analyze competitor titles with AI
 */
export const AnalyzeTitlesBody = zod.object({
  titles: zod.array(zod.string()),
  outputMode: zod.enum(["docs", "text"]),
  customPrompt: zod.string().nullish(),
});

export const AnalyzeTitlesResponse = zod.object({
  id: zod.number(),
  analysis: zod.string(),
  outputMode: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Generate viral titles based on analysis
 */
export const GenerateTitlesBody = zod.object({
  analysis: zod.string(),
  niche: zod.string().nullish(),
  outputMode: zod.enum(["docs", "text"]),
  customGeneratePrompt: zod.string().nullish(),
});

export const GenerateTitlesResponse = zod.object({
  titles: zod.array(zod.string()),
  explanation: zod.string(),
});

/**
 * @summary Analyze a competitor script with AI
 */
export const AnalyzeScriptBody = zod.object({
  script: zod.string(),
  outputMode: zod.enum(["docs", "text"]),
});

export const AnalyzeScriptResponse = zod.object({
  id: zod.number(),
  analysis: zod.string(),
  outputMode: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Generate a full YouTube script
 */
export const GenerateScriptBody = zod.object({
  title: zod.string(),
  scriptAnalysis: zod.string().nullish(),
  targetWordCount: zod.number().nullish(),
  outputMode: zod.enum(["docs", "text"]),
});

export const GenerateScriptResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  script: zod.string(),
  wordCount: zod.number(),
  outputMode: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get saved title analyses
 */
export const GetTitleAnalysesResponse = zod.object({
  analyses: zod.array(
    zod.object({
      id: zod.number(),
      titles: zod.array(zod.string()),
      analysis: zod.string(),
      outputMode: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get saved script analyses
 */
export const GetScriptAnalysesResponse = zod.object({
  analyses: zod.array(
    zod.object({
      id: zod.number(),
      scriptPreview: zod.string(),
      analysis: zod.string(),
      outputMode: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get saved generated scripts
 */
export const GetGeneratedScriptsResponse = zod.object({
  scripts: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      script: zod.string(),
      wordCount: zod.number(),
      outputMode: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardStatsResponse = zod.object({
  totalVideos: zod.number(),
  totalCompetitors: zod.number(),
  totalTitleAnalyses: zod.number(),
  totalScriptAnalyses: zod.number(),
  totalGeneratedScripts: zod.number(),
  recentActivity: zod.array(
    zod.object({
      type: zod.string(),
      label: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});